server/requirements.txt
# Web Framework 
fastapi 
uvicorn[standard] 
 
# HTTP Requests 
requests 
certifi
 
# Firebase 
firebase-admin 
 
# Environment variables 
python-dotenv 
 
# Scheduler 
apscheduler 

# Image Processing
Pillow
pytesseract
